requirements.txt
python-binance
psycopg2-binary
pandas
flask
werkzeug==2.0.1
Flask-HTTPAuth
tqdm
TA-Lib
numpy
matplotlib
SQLAlchemy